Update script.js
This will print the results to the terminal using Node, and after doing this a few times, I get a lot of repetition. But  that's just because the variety of options isn't big enough to provide less repetition. 
So I'd say we just add more homes, names, and motos. But other than that this looks great. So I'll get on that.